Move L2VPN resources from ipam to vpn
