+++ netbox/ipam/api/nested_serializers.py
@@ -3,7 +3,6 @@ from rest_framework import serializers
 
 from ipam import models
 from netbox.api.serializers import WritableNestedSerializer
-from vpn.models import L2VPN, L2VPNTermination
 from .field_serializers import IPAddressField
 
 __all__ = [
@@ -14,8 +13,6 @@ __all__ = [
     'NestedFHRPGroupAssignmentSerializer',
     'NestedIPAddressSerializer',
     'NestedIPRangeSerializer',
-    'NestedL2VPNSerializer',
-    'NestedL2VPNTerminationSerializer',
     'NestedPrefixSerializer',
     'NestedRIRSerializer',
     'NestedRoleSerializer',
@@ -223,28 +220,3 @@ class NestedServiceSerializer(WritableNestedSerializer):
     class Meta:
         model = models.Service
         fields = ['id', 'url', 'display', 'name', 'protocol', 'ports']
-
-#
-# L2VPN
-#
-
-
-class NestedL2VPNSerializer(WritableNestedSerializer):
-    url = serializers.HyperlinkedIdentityField(view_name='ipam-api:l2vpn-detail')
-
-    class Meta:
-        model = L2VPN
-        fields = [
-            'id', 'url', 'display', 'identifier', 'name', 'slug', 'type'
-        ]
-
-
-class NestedL2VPNTerminationSerializer(WritableNestedSerializer):
-    url = serializers.HyperlinkedIdentityField(view_name='ipam-api:l2vpntermination-detail')
-    l2vpn = NestedL2VPNSerializer()
-
-    class Meta:
-        model = L2VPNTermination
-        fields = [
-            'id', 'url', 'display', 'l2vpn'
-        ]
