+++ netbox/ipam/choices.py
@@ -172,52 +172,3 @@ class ServiceProtocolChoices(ChoiceSet):
         (PROTOCOL_UDP, 'UDP'),
         (PROTOCOL_SCTP, 'SCTP'),
     )
-
-
-class L2VPNTypeChoices(ChoiceSet):
-    TYPE_VPLS = 'vpls'
-    TYPE_VPWS = 'vpws'
-    TYPE_EPL = 'epl'
-    TYPE_EVPL = 'evpl'
-    TYPE_EPLAN = 'ep-lan'
-    TYPE_EVPLAN = 'evp-lan'
-    TYPE_EPTREE = 'ep-tree'
-    TYPE_EVPTREE = 'evp-tree'
-    TYPE_VXLAN = 'vxlan'
-    TYPE_VXLAN_EVPN = 'vxlan-evpn'
-    TYPE_MPLS_EVPN = 'mpls-evpn'
-    TYPE_PBB_EVPN = 'pbb-evpn'
-
-    CHOICES = (
-        ('VPLS', (
-            (TYPE_VPWS, 'VPWS'),
-            (TYPE_VPLS, 'VPLS'),
-        )),
-        ('VXLAN', (
-            (TYPE_VXLAN, 'VXLAN'),
-            (TYPE_VXLAN_EVPN, 'VXLAN-EVPN'),
-        )),
-        ('L2VPN E-VPN', (
-            (TYPE_MPLS_EVPN, 'MPLS EVPN'),
-            (TYPE_PBB_EVPN, 'PBB EVPN'),
-        )),
-        ('E-Line', (
-            (TYPE_EPL, 'EPL'),
-            (TYPE_EVPL, 'EVPL'),
-        )),
-        ('E-LAN', (
-            (TYPE_EPLAN, 'Ethernet Private LAN'),
-            (TYPE_EVPLAN, 'Ethernet Virtual Private LAN'),
-        )),
-        ('E-Tree', (
-            (TYPE_EPTREE, 'Ethernet Private Tree'),
-            (TYPE_EVPTREE, 'Ethernet Virtual Private Tree'),
-        )),
-    )
-
-    P2P = (
-        TYPE_VPWS,
-        TYPE_EPL,
-        TYPE_EPLAN,
-        TYPE_EPTREE
-    )
