+++ netbox/ipam/constants.py
@@ -86,9 +86,3 @@ VLANGROUP_SCOPE_TYPES = (
 # 16-bit port number
 SERVICE_PORT_MIN = 1
 SERVICE_PORT_MAX = 65535
-
-L2VPN_ASSIGNMENT_MODELS = Q(
-    Q(app_label='dcim', model='interface') |
-    Q(app_label='ipam', model='vlan') |
-    Q(app_label='virtualization', model='vminterface')
-)
