+++ netbox/ipam/graphql/types.py
@@ -1,10 +1,8 @@
 import graphene
 
-from extras.graphql.mixins import ContactsMixin
 from ipam import filtersets, models
 from netbox.graphql.scalars import BigInt
 from netbox.graphql.types import BaseObjectType, OrganizationalObjectType, NetBoxObjectType
-from vpn.models import L2VPN, L2VPNTermination
 
 __all__ = (
     'ASNType',
@@ -14,8 +12,6 @@ __all__ = (
     'FHRPGroupAssignmentType',
     'IPAddressType',
     'IPRangeType',
-    'L2VPNType',
-    'L2VPNTerminationType',
     'PrefixType',
     'RIRType',
     'RoleType',
@@ -189,19 +185,3 @@ class VRFType(NetBoxObjectType):
         model = models.VRF
         fields = '__all__'
         filterset_class = filtersets.VRFFilterSet
-
-
-class L2VPNType(ContactsMixin, NetBoxObjectType):
-    class Meta:
-        model = L2VPN
-        fields = '__all__'
-        filtersets_class = filtersets.L2VPNFilterSet
-
-
-class L2VPNTerminationType(NetBoxObjectType):
-    assigned_object = graphene.Field('ipam.graphql.gfk_mixins.L2VPNAssignmentType')
-
-    class Meta:
-        model = L2VPNTermination
-        exclude = ('assigned_object_type', 'assigned_object_id')
-        filtersets_class = filtersets.L2VPNTerminationFilterSet
