+++ netbox/ipam/search.py
@@ -1,5 +1,4 @@
 from netbox.search import SearchIndex, register_search
-from vpn.models import L2VPN
 from . import models
 
 
@@ -68,18 +67,6 @@ class IPRangeIndex(SearchIndex):
         ('comments', 5000),
     )
     display_attrs = ('vrf', 'tenant', 'status', 'role', 'description')
-
-
-@register_search
-class L2VPNIndex(SearchIndex):
-    model = L2VPN
-    fields = (
-        ('name', 100),
-        ('slug', 110),
-        ('description', 500),
-        ('comments', 5000),
-    )
-    display_attrs = ('type', 'identifier', 'tenant', 'description')
 
 
 @register_search
