+++ netbox/ipam/tables/__init__.py
@@ -1,7 +1,6 @@
 from .asn import *
 from .fhrp import *
 from .ip import *
-from .l2vpn import *
 from .services import *
 from .vlans import *
 from .vrfs import *
