+++ netbox/ipam/tests/test_views.py
@@ -9,8 +9,7 @@ from dcim.models import Device, DeviceRole, DeviceType, Manufacturer, Site, Inte
 from ipam.choices import *
 from ipam.models import *
 from tenancy.models import Tenant
-from utilities.testing import ViewTestCases, create_test_device, create_tags
-from vpn.models import L2VPN, L2VPNTermination
+from utilities.testing import ViewTestCases, create_tags
 
 
 class ASNRangeTestCase(ViewTestCases.PrimaryObjectViewTestCase):
@@ -987,142 +986,3 @@ class ServiceTestCase(ViewTestCases.PrimaryObjectViewTestCase):
         self.assertEqual(instance.protocol, service_template.protocol)
         self.assertEqual(instance.ports, service_template.ports)
         self.assertEqual(instance.description, service_template.description)
-
-
-class L2VPNTestCase(ViewTestCases.PrimaryObjectViewTestCase):
-    model = L2VPN
-
-    @classmethod
-    def setUpTestData(cls):
-        rts = (
-            RouteTarget(name='64534:123'),
-            RouteTarget(name='64534:321')
-        )
-        RouteTarget.objects.bulk_create(rts)
-
-        l2vpns = (
-            L2VPN(name='L2VPN 1', slug='l2vpn-1', type=L2VPNTypeChoices.TYPE_VXLAN, identifier='650001'),
-            L2VPN(name='L2VPN 2', slug='l2vpn-2', type=L2VPNTypeChoices.TYPE_VXLAN, identifier='650002'),
-            L2VPN(name='L2VPN 3', slug='l2vpn-3', type=L2VPNTypeChoices.TYPE_VXLAN, identifier='650003')
-        )
-        L2VPN.objects.bulk_create(l2vpns)
-
-        cls.csv_data = (
-            'name,slug,type,identifier',
-            'L2VPN 5,l2vpn-5,vxlan,456',
-            'L2VPN 6,l2vpn-6,vxlan,444',
-        )
-
-        cls.csv_update_data = (
-            'id,name,description',
-            f'{l2vpns[0].pk},L2VPN 7,New description 7',
-            f'{l2vpns[1].pk},L2VPN 8,New description 8',
-        )
-
-        cls.bulk_edit_data = {
-            'description': 'New Description',
-        }
-
-        cls.form_data = {
-            'name': 'L2VPN 8',
-            'slug': 'l2vpn-8',
-            'type': L2VPNTypeChoices.TYPE_VXLAN,
-            'identifier': 123,
-            'description': 'Description',
-            'import_targets': [rts[0].pk],
-            'export_targets': [rts[1].pk]
-        }
-
-
-class L2VPNTerminationTestCase(
-        ViewTestCases.GetObjectViewTestCase,
-        ViewTestCases.GetObjectChangelogViewTestCase,
-        ViewTestCases.CreateObjectViewTestCase,
-        ViewTestCases.EditObjectViewTestCase,
-        ViewTestCases.DeleteObjectViewTestCase,
-        ViewTestCases.ListObjectsViewTestCase,
-        ViewTestCases.BulkImportObjectsViewTestCase,
-        ViewTestCases.BulkDeleteObjectsViewTestCase,
-):
-
-    model = L2VPNTermination
-
-    @classmethod
-    def setUpTestData(cls):
-        device = create_test_device('Device 1')
-        interface = Interface.objects.create(name='Interface 1', device=device, type='1000baset')
-        l2vpns = (
-            L2VPN(name='L2VPN 1', slug='l2vpn-1', type=L2VPNTypeChoices.TYPE_VXLAN, identifier=650001),
-            L2VPN(name='L2VPN 2', slug='l2vpn-2', type=L2VPNTypeChoices.TYPE_VXLAN, identifier=650002),
-        )
-        L2VPN.objects.bulk_create(l2vpns)
-
-        vlans = (
-            VLAN(name='Vlan 1', vid=1001),
-            VLAN(name='Vlan 2', vid=1002),
-            VLAN(name='Vlan 3', vid=1003),
-            VLAN(name='Vlan 4', vid=1004),
-            VLAN(name='Vlan 5', vid=1005),
-            VLAN(name='Vlan 6', vid=1006)
-        )
-        VLAN.objects.bulk_create(vlans)
-
-        terminations = (
-            L2VPNTermination(l2vpn=l2vpns[0], assigned_object=vlans[0]),
-            L2VPNTermination(l2vpn=l2vpns[0], assigned_object=vlans[1]),
-            L2VPNTermination(l2vpn=l2vpns[0], assigned_object=vlans[2])
-        )
-        L2VPNTermination.objects.bulk_create(terminations)
-
-        cls.form_data = {
-            'l2vpn': l2vpns[0].pk,
-            'device': device.pk,
-            'interface': interface.pk,
-        }
-
-        cls.csv_data = (
-            "l2vpn,vlan",
-            "L2VPN 1,Vlan 4",
-            "L2VPN 1,Vlan 5",
-            "L2VPN 1,Vlan 6",
-        )
-
-        cls.csv_update_data = (
-            f"id,l2vpn",
-            f"{terminations[0].pk},{l2vpns[0].name}",
-            f"{terminations[1].pk},{l2vpns[0].name}",
-            f"{terminations[2].pk},{l2vpns[0].name}",
-        )
-
-        cls.bulk_edit_data = {}
-
-    # TODO: Fix L2VPNTerminationImportForm validation to support bulk updates
-    def test_bulk_update_objects_with_permission(self):
-        pass
-
-    #
-    # Custom assertions
-    #
-
-    # TODO: Remove this
-    def assertInstanceEqual(self, instance, data, exclude=None, api=False):
-        """
-        Override parent
-        """
-        if exclude is None:
-            exclude = []
-
-        fields = [k for k in data.keys() if k not in exclude]
-        model_dict = self.model_to_dict(instance, fields=fields, api=api)
-
-        # Omit any dictionary keys which are not instance attributes or have been excluded
-        relevant_data = {
-            k: v for k, v in data.items() if hasattr(instance, k) and k not in exclude
-        }
-
-        # Handle relations on the model
-        for k, v in model_dict.items():
-            if isinstance(v, object) and hasattr(v, 'first'):
-                model_dict[k] = v.first().pk
-
-        self.assertDictEqual(model_dict, relevant_data)
