+++ netbox/ipam/urls.py
@@ -131,20 +131,4 @@ urlpatterns = [
     path('services/edit/', views.ServiceBulkEditView.as_view(), name='service_bulk_edit'),
     path('services/delete/', views.ServiceBulkDeleteView.as_view(), name='service_bulk_delete'),
     path('services/<int:pk>/', include(get_model_urls('ipam', 'service'))),
-
-    # L2VPN
-    path('l2vpns/', views.L2VPNListView.as_view(), name='l2vpn_list'),
-    path('l2vpns/add/', views.L2VPNEditView.as_view(), name='l2vpn_add'),
-    path('l2vpns/import/', views.L2VPNBulkImportView.as_view(), name='l2vpn_import'),
-    path('l2vpns/edit/', views.L2VPNBulkEditView.as_view(), name='l2vpn_bulk_edit'),
-    path('l2vpns/delete/', views.L2VPNBulkDeleteView.as_view(), name='l2vpn_bulk_delete'),
-    path('l2vpns/<int:pk>/', include(get_model_urls('ipam', 'l2vpn'))),
-
-    # L2VPN terminations
-    path('l2vpn-terminations/', views.L2VPNTerminationListView.as_view(), name='l2vpntermination_list'),
-    path('l2vpn-terminations/add/', views.L2VPNTerminationEditView.as_view(), name='l2vpntermination_add'),
-    path('l2vpn-terminations/import/', views.L2VPNTerminationBulkImportView.as_view(), name='l2vpntermination_import'),
-    path('l2vpn-terminations/edit/', views.L2VPNTerminationBulkEditView.as_view(), name='l2vpntermination_bulk_edit'),
-    path('l2vpn-terminations/delete/', views.L2VPNTerminationBulkDeleteView.as_view(), name='l2vpntermination_bulk_delete'),
-    path('l2vpn-terminations/<int:pk>/', include(get_model_urls('ipam', 'l2vpntermination'))),
 ]
