+++ netbox/templates/vpn/l2vpn.html
@@ -34,7 +34,7 @@
         </table>
       </div>
     </div>
-    {% include 'inc/panels/tags.html' with tags=object.tags.all url='ipam:l2vpn_list' %}
+    {% include 'inc/panels/tags.html' with tags=object.tags.all url='vpn:l2vpn_list' %}
     {% plugin_left_page object %}
 	</div>
 	<div class="col col-md-6">
@@ -56,12 +56,12 @@
     <div class="card">
       <h5 class="card-header">{% trans "Terminations" %}</h5>
       <div class="card-body htmx-container table-responsive"
-        hx-get="{% url 'ipam:l2vpntermination_list' %}?l2vpn_id={{ object.pk }}"
+        hx-get="{% url 'vpn:l2vpntermination_list' %}?l2vpn_id={{ object.pk }}"
         hx-trigger="load"
       ></div>
-      {% if perms.ipam.add_l2vpntermination %}
+      {% if perms.vpn.add_l2vpntermination %}
         <div class="card-footer text-end noprint">
-          <a href="{% url 'ipam:l2vpntermination_add' %}?l2vpn={{ object.pk }}&return_url={{ object.get_absolute_url }}" class="btn btn-primary btn-sm{% if not object.can_add_termination %} disabled" aria-disabled="true{% endif %}">
+          <a href="{% url 'vpn:l2vpntermination_add' %}?l2vpn={{ object.pk }}&return_url={{ object.get_absolute_url }}" class="btn btn-primary btn-sm{% if not object.can_add_termination %} disabled" aria-disabled="true{% endif %}">
             <i class="mdi mdi-plus-thick" aria-hidden="true"></i> {% trans "Add a Termination" %}
           </a>
         </div>
