+++ netbox/templates/vpn/l2vpntermination.html
@@ -25,7 +25,7 @@
 	</div>
 	<div class="col col-md-6">
         {% include 'inc/panels/custom_fields.html' %}
-        {% include 'inc/panels/tags.html' with tags=object.tags.all url='ipam:l2vpntermination_list' %}
+        {% include 'inc/panels/tags.html' with tags=object.tags.all url='vpn:l2vpntermination_list' %}
     </div>
 </div>
