+++ netbox/templates/vpn/l2vpntermination_edit.html
