+++ netbox/vpn/forms/bulk_import.py
@@ -1,7 +1,8 @@
+from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
 from dcim.models import Device, Interface
-from ipam.models import IPAddress
+from ipam.models import IPAddress, VLAN
 from netbox.forms import NetBoxModelImportForm
 from tenancy.models import Tenant
 from utilities.forms.fields import CSVChoiceField, CSVModelChoiceField, CSVModelMultipleChoiceField
@@ -15,6 +16,8 @@ __all__ = (
     'IPSecPolicyImportForm',
     'IPSecProfileImportForm',
     'IPSecProposalImportForm',
+    'L2VPNImportForm',
+    'L2VPNTerminationImportForm',
     'TunnelImportForm',
     'TunnelTerminationImportForm',
 )
@@ -228,3 +231,92 @@ class IPSecProfileImportForm(NetBoxModelImportForm):
         fields = (
             'name', 'mode', 'ike_policy', 'ipsec_policy', 'description', 'comments', 'tags',
         )
+
+
+class L2VPNImportForm(NetBoxModelImportForm):
+    tenant = CSVModelChoiceField(
+        label=_('Tenant'),
+        queryset=Tenant.objects.all(),
+        required=False,
+        to_field_name='name',
+    )
+    type = CSVChoiceField(
+        label=_('Type'),
+        choices=L2VPNTypeChoices,
+        help_text=_('L2VPN type')
+    )
+
+    class Meta:
+        model = L2VPN
+        fields = ('identifier', 'name', 'slug', 'tenant', 'type', 'description',
+                  'comments', 'tags')
+
+
+class L2VPNTerminationImportForm(NetBoxModelImportForm):
+    l2vpn = CSVModelChoiceField(
+        queryset=L2VPN.objects.all(),
+        required=True,
+        to_field_name='name',
+        label=_('L2VPN'),
+    )
+    device = CSVModelChoiceField(
+        label=_('Device'),
+        queryset=Device.objects.all(),
+        required=False,
+        to_field_name='name',
+        help_text=_('Parent device (for interface)')
+    )
+    virtual_machine = CSVModelChoiceField(
+        label=_('Virtual machine'),
+        queryset=VirtualMachine.objects.all(),
+        required=False,
+        to_field_name='name',
+        help_text=_('Parent virtual machine (for interface)')
+    )
+    interface = CSVModelChoiceField(
+        label=_('Interface'),
+        queryset=Interface.objects.none(),  # Can also refer to VMInterface
+        required=False,
+        to_field_name='name',
+        help_text=_('Assigned interface (device or VM)')
+    )
+    vlan = CSVModelChoiceField(
+        label=_('VLAN'),
+        queryset=VLAN.objects.all(),
+        required=False,
+        to_field_name='name',
+        help_text=_('Assigned VLAN')
+    )
+
+    class Meta:
+        model = L2VPNTermination
+        fields = ('l2vpn', 'device', 'virtual_machine', 'interface', 'vlan', 'tags')
+
+    def __init__(self, data=None, *args, **kwargs):
+        super().__init__(data, *args, **kwargs)
+
+        if data:
+
+            # Limit interface queryset by device or VM
+            if data.get('device'):
+                self.fields['interface'].queryset = Interface.objects.filter(
+                    **{f"device__{self.fields['device'].to_field_name}": data['device']}
+                )
+            elif data.get('virtual_machine'):
+                self.fields['interface'].queryset = VMInterface.objects.filter(
+                    **{f"virtual_machine__{self.fields['virtual_machine'].to_field_name}": data['virtual_machine']}
+                )
+
+    def clean(self):
+        super().clean()
+
+        if self.cleaned_data.get('device') and self.cleaned_data.get('virtual_machine'):
+            raise ValidationError(_('Cannot import device and VM interface terminations simultaneously.'))
+        if not self.instance and not (self.cleaned_data.get('interface') or self.cleaned_data.get('vlan')):
+            raise ValidationError(_('Each termination must specify either an interface or a VLAN.'))
+        if self.cleaned_data.get('interface') and self.cleaned_data.get('vlan'):
+            raise ValidationError(_('Cannot assign both an interface and a VLAN.'))
+
+        # if this is an update we might not have interface or vlan in the form data
+        if self.cleaned_data.get('interface') or self.cleaned_data.get('vlan'):
+            self.instance.assigned_object = self.cleaned_data.get('interface') or self.cleaned_data.get('vlan')
