+++ netbox/vpn/tables/__init__.py
@@ -1,0 +1,3 @@
+from .crypto import *
+from .l2vpn import *
+from .tunnels import *
